+++ views/chatbotPage.ejs
@@ -1,7 +1,7 @@
 <%- include("templates/headerTemplate")%>
 
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
-<script src="./chat2.js"></script>
+<!-- <script src="/chat2.js"></script> -->
     <div
         class="max-w-[450px] w-full min-h-svh mx-auto py-10 px-[35px] pb-[70px] bg-gray-900 flex flex-col items-center justify-between space-y-5 text-sky-100 text-md font-medium">
         <!-- Page title -->
@@ -57,7 +57,7 @@
             </div>
 
             <!-- Text box -->
-            <form  id="chatInput">
+            <form action="/chatbot" method="post" id="chatInput">
                 <input name="inputBox" type="text" id="chatbotTextBox" contenteditable rows="1" style="scrollbar-width: none;"
                     class="w-full min-h-[25px] max-h-[100px] bg-gray-700 resize-none overflow-y-scroll text-sky-100 empty:before:content-['Ask_Kate...'] focus:outline-none">
                 </input>
